Address sandbox policy review feedback
